+++ keyboards/barleycorn/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "Houndstooth",
+    "keyboard_name": "Barleycorn",
     "maintainer": "Yiancar",
     "width": 19.5,
     "height": 5,
