+++ keyboards/barleycorn/keymaps/default/readme.md
@@ -1,4 +1,4 @@
-The default keymap for Houndstooth
+The default keymap for Barleycorn
 ==================================
 
 ![Layer 0](https://i.imgur.com/cU9lzz3.png)
