+++ keyboards/barleycorn/keymaps/iso/readme.md
@@ -1,4 +1,4 @@
-The default keymap for ISO Houndstooth
+The default keymap for ISO Barleycorn
 ======================================
 
 ![Layer 0](https://i.imgur.com/llZlVHt.png)
